* Introduced new BBitmap flag B_BITMAP_SCALE_BILINEAR. * When drawing BBitmaps with scaling in the app_server, use a bilinear   filter when a bitmap has this flag set. (Hope nobody objects, otherwise   I can revert or improve this. Performance can certainly be improved, since   the AGG implementation is too generic. But that goes for the nearest   neighbor implementation as well.) * Flags are uint32, fix app_server side code to declare them correctly. Use   appropriate link methods in BBitmap and ServerApp. * Enable the BeOS compatibility mode for B_RGB32 (works just like B_RGBA32   in B_OP_ALPHA mode).
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@26649 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/interface/Bitmap.cpp
@@ -997,7 +997,7 @@ BBitmap::_InitObject(BRect bounds, color_space colorSpace, uint32 flags,
 			link.StartMessage(AS_CREATE_BITMAP);
 			link.Attach<BRect>(bounds);
 			link.Attach<color_space>(colorSpace);
-			link.Attach<int32>((int32)flags);
+			link.Attach<uint32>(flags);
 			link.Attach<int32>(bytesPerRow);
 			link.Attach<int32>(screenID.id);
